Add cron task to fix multiple completed 3 scores, score is now int64

+++ fix_score_duplicates.go
@@ -10,7 +10,7 @@ type score struct {
 	id         int
 	beatmapMD5 string
 	userid     int
-	score      int
+	score      int64
 	maxCombo   int
 	mods       int
 	playMode   int
